reorganize the current code
The generators are now all in their own namespace, in
ConstantSetpointGenerators, as they are completely
independent of the type of control chain we use.

The direct* controllers have been superseded by the StablePitchRoll
control strategy, where we assume that pitch and roll are always zero
(or at least negligible) and therefore aligned == body. It is a bit
more complex than the direct controls, but not overly so. At least,
they do not require anything more on the sensing side (only heading for
yaw and depth for Z)