auv_control: fix dynamic determination of the expected inputs
Since we attempt to dynamically reconfigure the auv_control components
based on the network shape, we have to account for control cascades.

If there is two PID controllers in a chain
   PID0 -> PID1

Then a disconnection on the inputs of PID0 will change its output domain
and affect the expected inputs on PID1. This has to be propagated properly.

The commit achieves this by using a task-level #full_output_domain method
instead of the domain defined by services if such a method exists.

Note that at the level of network generation, this is not a problem
as at this stage services exactly represent the actual input/output.
The issue this commit attempts to fix is caused by the usage of
already-running components whose services are a superset of the actual
services required in the network.